Refactor books component
