+++ src/content/books/books.json
@@ -1,6 +1,12 @@
 {
     "books": [
         {
+            "title": "The Rosie Project",
+            "coverImageName": "rosieproject",
+            "author": "Graeme Simsion",
+            "year": 2020
+        },
+        {
             "title": "Happy",
             "coverImageName": "happy",
             "author": "Derren Brown",
@@ -11,12 +17,6 @@
             "coverImageName": "makingofamanager",
             "author": "Julie Zhuo",
             "year": 2019
-        },
-        {
-            "title": "The Rosie Project",
-            "coverImageName": "rosieproject",
-            "author": "Graeme Simsion",
-            "year": 2020
         },
         {
             "title": "Sapiens",
